added vscode tasks for code formating, building and debugging

+++ start/scripts/create_package.sh
@@ -1,5 +1,15 @@
 #!/bin/bash -x
 
+#
+# first parameter of this script is the build type which is either Debug or Release
+#
+if [ -z "$1" ]
+  then
+    echo "No argument supplied"
+    exit 1
+fi
+build_type_argument=$1
+
 # Clean build folder
 rm -rf build
 
@@ -7,14 +17,14 @@ rm -rf build
 conan profile detect --force
 
 # Install conan package
-conan install . --output-folder=build --build=missing
+conan install . --output-folder=build --build=missing --settings=build_type=$build_type_argument
 
 # Generate build system
 cd build
-cmake .. -DCMAKE_TOOLCHAIN_FILE=conan_toolchain.cmake -DCMAKE_BUILD_TYPE=Release
+cmake .. -DCMAKE_TOOLCHAIN_FILE=conan_toolchain.cmake -DCMAKE_BUILD_TYPE=$build_type_argument
 
 # Build
-cmake --build . --config Release
+cmake --build . --config $build_type_argument
 
 # Run test scripts
 ctest .
